+++ pipe_type.h
@@ -2,10 +2,10 @@
 
 class PIPE_TYPE
 {
-	float capacity; // przepustowoœæ
-	float price; // cena w jednostkach waluty / jednostkê odleg³oœci
+	double capacity; // przepustowoœæ
+	double price; // cena w jednostkach waluty / jednostkê odleg³oœci
 
-	PIPE_TYPE(float capacity, float price) : capacity(capacity), price(price) {}
+	PIPE_TYPE(double capacity, double price) : capacity(capacity), price(price) {}
 	PIPE_TYPE(const PIPE_TYPE&) = delete;
 	PIPE_TYPE& operator= (const PIPE_TYPE&) = delete;
